requirements.txt
flask==3.0.0
anthropic==0.39.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
requests==2.31.0
gunicorn==21.2.0
